增加spring-cloud-config 集中配置总件和spring-cloud-bus消息总线总件 前者为把配置放在云端方便修改配置  后者为启动项目的情况下更改配置生效且不用启动

+++ tensquare_base/src/main/java/com/tensquare/base/controller/LabelController.java
@@ -5,23 +5,34 @@ import com.tensquare.base.service.LabelService;
 import entity.PageResult;
 import entity.Result;
 import entity.StatusCode;
+import io.jsonwebtoken.Claims;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.cloud.context.config.annotation.RefreshScope;
 import org.springframework.data.domain.Page;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 
 @RestController
 @CrossOrigin
 @RequestMapping("/label")
+@RefreshScope      //自定义注解要加入这个可以使消息队列监听最新配置文件并获取
 public class LabelController {
 
     @Autowired
     private LabelService labelService;
 
+    @Autowired
+    private HttpServletRequest request;
+
+    @Value("${ip}") //自定义注解
+    private String ip;
 
     @GetMapping
     public Result findAll(){
+        System.out.println("ip为 "+ip);
         return new Result(true, StatusCode.OK,"查询成功",labelService.findAll());
     }
 
@@ -33,18 +44,30 @@ public class LabelController {
 
     @PostMapping
     public Result save(@RequestBody Label label){
+        Claims claims =(Claims)request.getAttribute("claims_user");
+        if (claims==null){
+            return  new Result(false,StatusCode.ACCESSERROR,"权限不足");
+        }
         labelService.save(label);
         return new Result(true, StatusCode.OK,"添加成功");
     }
 
     @DeleteMapping("/{labelId}")
     public Result deleteById(@PathVariable String labelId){
+        Claims claims =(Claims)request.getAttribute("claims_user");
+        if (claims==null){
+            return  new Result(false,StatusCode.ACCESSERROR,"权限不足");
+        }
         labelService.deleteById(labelId);
         return new Result(true, StatusCode.OK,"删除成功");
     }
 
     @PutMapping("/{labelId}")
     public Result  update(@PathVariable String labelId,@RequestBody Label label){
+        Claims claims =(Claims)request.getAttribute("claims_user");
+        if (claims==null){
+            return  new Result(false,StatusCode.ACCESSERROR,"权限不足");
+        }
         label.setId(labelId);
         labelService.update(label);
         return new Result(true, StatusCode.OK,"修改成功");
